routing: Support alternatives for single-mode routing calculations
This requires to pass down the trip routing query attributes all the way
down to the OSRM call. It is added as an optional parameter at the end
of the `route` and other calculation functions, as well as in the socket
route for routing calculation.

If the `withAlternatives` attribute is `true` , the osrm backend call
sets the alternatives attribute to `true`.