utils: Add a `BaseObjectWithNumericId` abstract class for base objects
This class will replace the GenericObject class when the objects migrate
to having numeric IDs instead of uuid primary key.

The `BaseAttributesWithNumericId` type contains only the fields that are
common to all classes, not the optional ones, that should be added
independently on the concrete classes.

This greatly limits the functionalities that were available in the
GenericObject class, the `_attributes` is now a private field and the
`attributes` getter is readonly. The attributes can only be set through
the `set` method which is correctly typed to make sure `path` and
`value` have the expected values. Also, it is not possible to update an
object that is frozen, the `set` and `mergeAttributes` methods will
throw an error in those cases.