db: Add a `transaction` option to `exists` and `read` functions
If a read/exists call is done within a transaction on an element that
was added/modified during the transaction, if the transaction parameter
is not set, the object will not be correctly read.

This lets elements calls inside transaction to read the data as it is in
the transaction.